Major Updates (#179)
* [N/A] Update ACF to latest

* Adds media proxy functionality

Enables proxying media files (uploads) from another domain.

This feature allows users to specify a proxy domain during project creation,
which will then be used to serve media files.
This is achieved by adding an nginx configuration file and updating the
project creation script to prompt the user for the proxy domain and
configure the nginx proxy accordingly.

Addresses the need to serve media files from a different domain
than the main application, e.g., when migrating a site but retaining the
existing media URLs.

* [N/A] Better Media Proxy rename

* [N/A] Theme JSON updates

* Enables live reload and configures polling

Configures live reload to watch PHP files for changes and updates the browser.

Sets up polling for file watching, which helps in environments where the file system doesn't notify Vite of changes. This is configured with an interval of 1 second.

* [#177] Fix Skip Cropping bug

* [N/A] Some block cleanup

* [N/A] Added branding step during setup

* Removes the inline styling in the footer template.

* Version bump to 0.1.1 and changelog update

* Accidentally updated the version/changelog in the wrong spot.